Adds local and remote modules
